# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Il magazzino del man
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /proposals:
    get:
      operationId: listProposals
      tags: [proposals]
      summary: List all item proposals (admin only)
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, accepted, rejected]
      responses:
        "200":
          description: List of proposals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proposal"
    post:
      operationId: createProposal
      tags: [proposals]
      summary: Submit a new item proposal (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalInput"
      responses:
        "201":
          description: Created proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
  /proposals/{id}:
    get:
      operationId: getProposal
      tags: [proposals]
      summary: Get a single proposal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
  /proposals/{id}/accept:
    post:
      operationId: acceptProposal
      tags: [proposals]
      summary: Accept a proposal with commission (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommissionInput"
      responses:
        "200":
          description: Accepted proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
  /proposals/{id}/reject:
    post:
      operationId: rejectProposal
      tags: [proposals]
      summary: Reject a proposal (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
...
components:
  schemas:
...
    Proposal:
      type: object
      required: [id, itemName, category, desiredPayout, status, submitterEmail, createdAt]
      properties:
        id:
          type: integer
        itemName:
          type: string
        category:
          type: string
        desiredPayout:
          type: number
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, accepted, rejected]
        submitterEmail:
          type: string
        submitterName:
          type: ["string", "null"]
        commissionType:
          type: ["string", "null"]
          enum: [percentage, fixed, null]
        commissionValue:
          type: ["number", "null"]
        netProfit:
          type: ["number", "null"]
        adminNotes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ProposalInput:
      type: object
      required: [itemName, category, desiredPayout, submitterEmail]
      properties:
        itemName:
          type: string
        category:
          type: string
        desiredPayout:
          type: number
        imageUrl:
          type: string
        submitterEmail:
          type: string
        submitterName:
          type: string
    CommissionInput:
      type: object
      required: [commissionType, commissionValue]
      properties:
        commissionType:
          type: string
          enum: [percentage, fixed]
        commissionValue:
          type: number
        adminNotes:
          type: string
...